+++ flutter/lib/l10n/arb/intl_ru_RU.arb
@@ -605,9 +605,9 @@
     "placeholders": {
     }
   },
-  "reviewLearning": "Просмотр",
-  "@reviewLearning": {
-    "description": "Review Learning Label",
+  "viewLearning": "Просмотр",
+  "@viewLearning": {
+    "description": "View Learning Label",
     "type": "text",
     "placeholders": {
     }
