+++ flutter/lib/l10n/messages_en_US.dart
@@ -139,7 +139,6 @@ class MessageLookup extends MessageLookupByLibrary {
         "profileTooltip": MessageLookupByLibrary.simpleMessage("Profile"),
         "reversedCardLabel":
             MessageLookupByLibrary.simpleMessage("Add reversed card"),
-        "reviewLearning": MessageLookupByLibrary.simpleMessage("Review"),
         "save": MessageLookupByLibrary.simpleMessage("Save"),
         "saveChangesQuestion": MessageLookupByLibrary.simpleMessage(
             "Do you want to save changes?"),
@@ -160,6 +159,7 @@ class MessageLookup extends MessageLookupByLibrary {
         "supportDevelopment": MessageLookupByLibrary.simpleMessage(
             "Please tell us what we can do to make your experience with Delern better!\n\nIf you have any questions or suggestions please contact us:\n[delern@dasfoo.org](mailto:delern@dasfoo.org)\n\nFollow latest news on:\n\n- [Facebook](https://fb.me/das.delern)\n- [Twitter](https://twitter.com/dasdelern)\n- [VK](https://vk.com/delern)\n\nTo see the source code for this app, please visit the [Delern GitHub repo](https://github.com/dasfoo/delern).\n      "),
         "swissDeckType": MessageLookupByLibrary.simpleMessage("Swiss"),
+        "viewLearning": MessageLookupByLibrary.simpleMessage("View"),
         "watchedCards": m4,
         "whoHasAccessLabel":
             MessageLookupByLibrary.simpleMessage("Who has access"),
