+++ flutter/lib/views/decks_list/decks_list.dart
@@ -11,7 +11,7 @@ import 'package:delern_flutter/view_models/decks_list_bloc.dart';
 import 'package:delern_flutter/views/card_create_update/card_create_update.dart';
 import 'package:delern_flutter/views/cards_interval_learning/cards_interval_learning.dart';
 import 'package:delern_flutter/views/cards_list/cards_list.dart';
-import 'package:delern_flutter/views/cards_review_learning/cards_review_learning.dart';
+import 'package:delern_flutter/views/cards_view_learning/cards_view_learning.dart';
 import 'package:delern_flutter/views/decks_list/create_deck_widget.dart';
 import 'package:delern_flutter/views/decks_list/deck_menu.dart';
 import 'package:delern_flutter/views/decks_list/navigation_drawer.dart';
@@ -218,12 +218,12 @@ class DeckListItemWidget extends StatelessWidget {
                           },
                         ),
                         LearningMethodWidget(
-                          name: localizations.of(context).reviewLearning,
+                          name: localizations.of(context).viewLearning,
                           icon: Icons.remove_red_eye,
                           onTap: () async {
                             // Close dialog
                             Navigator.pop(context);
-                            await _learnCardsReview(context);
+                            await _learnCardsView(context);
                           },
                         ),
                       ],
@@ -254,11 +254,11 @@ class DeckListItemWidget extends StatelessWidget {
     }
   }
 
-  Future<void> _learnCardsReview(BuildContext context) async {
+  Future<void> _learnCardsView(BuildContext context) async {
     final anyCardsShown = await Navigator.push(
         context,
         MaterialPageRoute(
-          settings: const RouteSettings(name: '/decks/learn-review'),
+          settings: const RouteSettings(name: '/decks/learn-view'),
           builder: (context) => CardsReviewLearning(deck: deck),
         ));
     if (anyCardsShown == false) {
